Move StubValidationSettings to Config plugin
