+++ npm-packages/meteor-vite/src/MeteorViteConfig.ts
@@ -1,4 +1,5 @@
 import { ResolvedConfig } from 'vite';
+import { StubValidationSettings } from './plugin/Config';
 
 export declare interface MeteorViteConfig extends ResolvedConfig {
     meteor?: {
@@ -19,29 +20,3 @@ export declare interface MeteorViteConfig extends ResolvedConfig {
     };
 }
 
-export interface StubValidationSettings {
-    /**
-     * list of packages to ignore export validation for.
-     * @example
-     * { ignorePackages: ['ostrio:cookies', 'test:ts-modules', ...] }
-     */
-    ignorePackages?: string[];
-    
-    /**
-     * Will only emit warnings in the console instead of throwing an exception that may prevent the client app
-     * from loading.
-     * @default true
-     */
-    warnOnly?: boolean;
-    
-    /**
-     * Whether to completely disable stub validation feature for Meteor-Vite.
-     *
-     * Tip:
-     * You can use a conditional Vite configuration to enable/disable this for your production build
-     * {@link https://vitejs.dev/config/#conditional-config}
-     *
-     * @default false
-     */
-    disabled?: boolean;
-}
