+++ npm-packages/meteor-vite/src/meteor/package/StubTemplate.ts
@@ -1,5 +1,5 @@
-import { StubValidationSettings } from '../../MeteorViteConfig';
 import { StubValidatorOptions } from '../../client/ValidateStub';
+import { StubValidationSettings } from '../../plugin/Config';
 import MeteorPackage from './components/MeteorPackage';
 
 export const METEOR_STUB_KEY = `m2`;
